refactor: view초기화 부분 view디렉토리로 이동 및 파일 분할
